bench: switch revm bench to 100 transfers
To better match the benchmarks in `revm` crate
